Move updatable file types to /models
